Feat: 훅 js => ts로 변경 및 개편
